+++ src/model/Model_Budget.cpp
@@ -149,11 +149,11 @@ void Model_Budget::getBudgetStats(
     }
 
     //Calculations
-    std::map<int, double> monthlyBudgetValue;
-    std::map<int, double> yearlyBudgetValue;
-    std::map<int, double> yearDeduction;
-    std::map<std::pair<int, int>, bool> isBudgeted;
-    std::map<int, int> budgetedMonths;
+    std::map<int64, double> monthlyBudgetValue;
+    std::map<int64, double> yearlyBudgetValue;
+    std::map<int64, double> yearDeduction;
+    std::map<std::pair<int, int64>, bool> isBudgeted;
+    std::map<int64, int> budgetedMonths;
     const wxString year = wxString::Format("%i", start_date.GetYear());
     int64 budgetYearID = Model_Budgetyear::instance().Get(year);
     for (const auto& budget : instance().find(BUDGETYEARID(budgetYearID)))
@@ -175,7 +175,7 @@ void Model_Budget::getBudgetStats(
         //fill with amount from monthly budgets first
         for (const auto& budget : instance().find(BUDGETYEARID(budgetYearID)))
         {
-            std::pair<int, int> month_categ = std::make_pair(month, budget.CATEGID);
+            std::pair<int, int64> month_categ = std::make_pair(month, budget.CATEGID);
             if (!isBudgeted[month_categ])
             {
                 isBudgeted[month_categ] = true;
@@ -216,7 +216,7 @@ void Model_Budget::getBudgetStats(
     }
     if (!groupByMonth)
     {
-        std::map<int, std::map<int,double> > yearlyBudgetStats;
+        std::map<int64, std::map<int,double> > yearlyBudgetStats;
         for (const auto& category : Model_Category::instance().all()) {
             yearlyBudgetStats[category.CATEGID][0] = 0.0;
         }
@@ -231,7 +231,7 @@ void Model_Budget::getBudgetStats(
 
 void Model_Budget::copyBudgetYear(int64 newYearID, int64 baseYearID)
 {
-    std::map<int, double> yearDeduction;
+    std::map<int64, double> yearDeduction;
     int budgetedMonths = 0;
     bool optionDeductMonthly = Option::instance().BudgetDeductMonthly();
     const wxString baseBudgetYearName = Model_Budgetyear::instance().get(baseYearID)->BUDGETYEARNAME;
@@ -244,7 +244,7 @@ void Model_Budget::copyBudgetYear(int64 newYearID, int64 baseYearID)
         for (int month = 0; month < 12; month++)
         {
             const wxString budgetYearMonth = wxString::Format("%s-%02d", newBudgetYearName.SubString(0,3), month + 1);
-            int budgetYearID = Model_Budgetyear::instance().Get(budgetYearMonth);
+            int64 budgetYearID = Model_Budgetyear::instance().Get(budgetYearMonth);
             Model_Budget::Data_Set monthlyBudgetData = instance().find(BUDGETYEARID(budgetYearID));
             if (!monthlyBudgetData.empty()) budgetedMonths++;
             //calculate deduction
