+++ src/reports/summarystocks.h
@@ -34,7 +34,7 @@ public:
 private:
     // structure for sorting of data
     struct data_holder { wxString name; wxString symbol; wxString date; double qty; double purchase; double current; double commission; double realgainloss; double unrealgainloss; double value; };
-    struct account_holder { int id; wxString name; std::vector<data_holder> data; double realgainloss; double unrealgainloss; double total; };
+    struct account_holder { int64 id; wxString name; std::vector<data_holder> data; double realgainloss; double unrealgainloss; double total; };
     std::vector<account_holder> m_stocks;
     double m_real_gain_loss_sum_total = 0.0;
     double m_unreal_gain_loss_sum_total = 0.0;
